+++ Tests/IntegrationTests.Shared/Person.cs
@@ -22,6 +22,7 @@ using System.Collections.Generic;
 
 namespace IntegrationTests
 {
+    [Preserve(AllMembers = true)]
     public class Person : RealmObject
     {
         // Automatically implemented (overridden) properties