+++ Tests/IntegrationTests.Shared/RelationshipTests.cs
@@ -28,9 +28,10 @@ using System.Diagnostics;
 // unlike the Cocoa definitions, we use Pascal casing for properties
 namespace IntegrationTests.Shared
 {
-    [TestFixture]
+    [TestFixture, Preserve(AllMembers = true)]
     public class RelationshipTests
     {
+        [Preserve(AllMembers = true)]
         class Dog : RealmObject
         {
             public string Name { get; set; }
@@ -39,6 +40,7 @@ namespace IntegrationTests.Shared
             //Owner Owner { get; set; }  will uncomment when verifying that we have back-links from ToMany relationships
         }
 
+        [Preserve(AllMembers = true)]
         class Owner : RealmObject
         {
             public string Name { get; set; }