+++ Tests/IntegrationTests.Shared/SimpleLINQtests.cs
@@ -25,6 +25,7 @@ using Realms;
 
 namespace IntegrationTests
 {
+    [TestFixture, Preserve(AllMembers = true)]
     class SimpleLINQtests : PeopleTestsBase
     {
         // see comment on base method why this isn't decorated with [SetUp]