+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 
 ### Minor Changes
 * Fixed a crash on iOS when creating many short-lived realms very rapidly in parallel (Issue #653)
+* Fixed a crash on iOS when creating many short-lived realms very rapidly in parallel (#653)
+* `RealmObject.IsValid` can be called to check if a managed object has been deleted
+* Accessing properties on invalid objects will throw an exception rather than crash with a segfault (#662)
 
 0.76.1 (2016-06-15)
 -------------------